+++ src/libs/model-runtime/utils/streams/vertex-ai.ts
@@ -31,13 +31,11 @@ const transformVertexAIStream = (
       { data: candidate.finishReason, id: context?.id, type: 'stop' },
       {
         data: {
-          // TODO: Google SDK 0.24.0 don't have promptTokensDetails types
-          inputImageTokens: (usage as any).promptTokensDetails?.find(
-            (i: any) => i.modality === 'IMAGE',
-          )?.tokenCount,
-          inputTextTokens: (usage as any).promptTokensDetails?.find(
-            (i: any) => i.modality === 'TEXT',
-          )?.tokenCount,
+          inputCachedTokens: usage.cachedContentTokenCount,
+          inputImageTokens: usage.promptTokensDetails?.find((i) => i.modality === 'IMAGE')
+            ?.tokenCount,
+          inputTextTokens: usage.promptTokensDetails?.find((i) => i.modality === 'TEXT')
+            ?.tokenCount,
           outputReasoningTokens,
           outputTextTokens,
           totalInputTokens: usage.promptTokenCount,
@@ -56,7 +54,7 @@ const transformVertexAIStream = (
     candidate.content.parts.length > 0
   ) {
     for (const part of candidate.content.parts) {
-      if (part && part.text && (part as any).thought === true) {
+      if (part && part.text && part.thought === true) {
         return { data: part.text, id: context.id, type: 'reasoning' };
       }
     }